+++ native/tests/test_host_open_default.c
@@ -1,7 +1,7 @@
 #define AIRUN_ALLOW_INTERNAL_UI_FALLBACK 1
 #define AIRUN_TEST_FAKE_OPEN_URL 1
 #define main airun_embedded_main_for_test
-#include "../../../AiCLI/native/airun.c"
+#include "../ailang_cli/ailang.c"
 #undef main
 
 #define CHECK(cond) \